+++ examples/book_subscription.py
@@ -8,7 +8,7 @@ from pydantic import ValidationError
 
 from architect_py.scalars import TradableProduct
 
-from .common import create_async_client
+from .common import connect_async_client
 
 buy_columns = "{:>15} {:>15} {:>15}"
 sell_columns = "{:<15} {:<15} {:<15}"
@@ -35,7 +35,7 @@ def print_book(book):
 
 
 async def main():
-    c: AsyncClient = await create_async_client()
+    c: AsyncClient = await connect_async_client()
     await c.grpc_client.change_channel("binance.marketdata.architect.co")
     symbol = TradableProduct("SOL-USDC BINANCE Perpetual/USDC Crypto")
     try: