+++ examples/candles.py
@@ -5,11 +5,11 @@ from architect_py.grpc_client.definitions import CandleWidth
 from architect_py.graphql_client.exceptions import GraphQLClientHttpError
 from architect_py.scalars import TradableProduct
 
-from .common import create_async_client
+from .common import connect_async_client
 
 
 async def main():
-    c: AsyncClient = await create_async_client()
+    c: AsyncClient = await connect_async_client()
     symbol = "ES 20250321 CME Future"
     quote = "USD"
     tradable_product = TradableProduct(symbol, quote)