+++ examples/stream_l1_marketdata.py
@@ -4,11 +4,11 @@ from uuid import UUID
 from architect_py.async_client import AsyncClient
 from architect_py.scalars import TradableProduct
 
-from .common import create_async_client
+from .common import connect_async_client
 
 
 async def main():
-    c: AsyncClient = await create_async_client()
+    c: AsyncClient = await connect_async_client()
 
     """
     CR acho: