+++ examples/trades.py
@@ -4,11 +4,11 @@ from architect_py.async_client import AsyncClient
 from architect_py.graphql_client.exceptions import GraphQLClientHttpError
 from architect_py.scalars import TradableProduct
 
-from .common import create_async_client
+from .common import connect_async_client
 
 
 async def main():
-    c: AsyncClient = await create_async_client()
+    c: AsyncClient = await connect_async_client()
     market_id = TradableProduct("BTC Crypto", "USD")
     try:
         async for trade in c.subscribe_trades_stream(market_id, venue="COINBASE"):