+++ examples/tutorial_async.py
@@ -4,11 +4,11 @@ from decimal import Decimal
 from architect_py.async_client import AsyncClient, OrderDir
 from architect_py.graphql_client.enums import OrderStatus
 from architect_py.scalars import TradableProduct
-from examples.common import create_async_client
+from examples.common import connect_async_client
 
 
 async def main():
-    c = await create_async_client()
+    c = await connect_async_client()
 
     market = TradableProduct("ES 20281215 CME Future/US")
     execution_venue = "CME"