+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 3.2.0
 Breaking:
-  - Changed the way AsyncClient is instantiated, from normal __init__ to create
-    - Can just change AsyncClient(args) -> AsyncClient.create(args)
+  - Changed the way AsyncClient is instantiated, from normal __init__ to 
+    - Can just change AsyncClient(args) -> AsyncClient.connect(args)
     - the port argument was removed, please remove it as an arg
   - changed the values of OrderDir to match Rust type
     class OrderDir(str, Enum):